Extract download_content.py from database.py
